+++ test/tab.firefox.test.ts
@@ -32,15 +32,15 @@ beforeEach(async () =>
 });
 
 
-afterEach(() =>
+afterEach(async () =>
 {
-  browser.close();
+  await browser.close();
 });
 
 
 test('screenshot', async () =>
 {
-  await tab.load("http://localhost:3001/ping");
+  await tab.load("http://localhost:3214/ping");
   await tab.screenshot();
 });
 
@@ -142,7 +142,7 @@ test.todo('solveReCaptchaV2');
 
 test('eval', async () =>
 {
-  await tab.load("http://localhost:3001/ping");
+  await tab.load("http://localhost:3214/ping");
 
   const body: string = await tab.eval(`() =>
   {